refactored package structure as well fixed redis spring connection
