+++ Dockerfile
@@ -5,7 +5,7 @@ WORKDIR /workspace/app
 COPY gradle gradle
 COPY build.gradle settings.gradle gradlew ./
 COPY src src
-
+RUN ls -a /root
 RUN target=/root/.gradle ./gradlew build -x test
 RUN mkdir -p build/libs/dependency && (cd build/libs/dependency; jar -xf ../*.jar)
 
@@ -17,4 +17,4 @@ COPY --from=build ${DEPENDENCY}/META-INF /app/META-INF
 COPY --from=build ${DEPENDENCY}/BOOT-INF/classes /app
 
 EXPOSE 23000
-ENTRYPOINT ["java","-cp","app:app/lib/*","com.TradeProject.TradeEngine.TradeEngine.TradeEngineApplication"]
+ENTRYPOINT ["java","-cp","app:app/lib/*","com.TradeProject.TradeEngine.TradeEngineApplication"]
