+++ docker-compose.staging.yml
@@ -5,3 +5,9 @@ services:
     image: elijahahianyo/trade-project:trade-engine-prod
     ports:
       - 23000:23000
+    networks:
+      - trade_project
+
+networks:
+  trade_project:
+    external: true
