+++ src/main/java/com/TradeProject/TradeEngine/HomeController.java
@@ -1,4 +1,4 @@
-package com.TradeProject.TradeEngine.TradeEngine;
+package com.TradeProject.TradeEngine;
 
 
 import org.springframework.stereotype.Controller;
