+++ src/main/resources/application.properties
@@ -1,2 +1,4 @@
 server.port = 23000
 redis.pubsub.channel.name=tradingEngine
+spring.redis.host=172.25.0.2
+spring.redis.port=6379
