+++ src/test/java/com/TradeProject/TradeEngine/TradeEngineApplicationTests.java
@@ -1,4 +1,4 @@
-package com.TradeProject.TradeEngine.TradeEngine;
+package com.TradeProject.TradeEngine;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
